requirements.txt
PyQt5
opencv-python
python-dotenv
google-cloud-vision
google-auth
groq
